Move SignalLevel color extension to rssi_view file and make it private
