+++ example/lib/util/signal_level.dart
@@ -1,22 +1,4 @@
-import 'package:flutter/material.dart';
-
 enum SignalLevel { high, medium, low, unknown }
-
-extension SignalLevelExtenstion on SignalLevel {
-  Color color() {
-    switch (this) {
-      case SignalLevel.high:
-        return Colors.green;
-      case SignalLevel.medium:
-        return Colors.orange;
-      case SignalLevel.low:
-        return Colors.red;
-      case SignalLevel.unknown:
-      default:
-        return Colors.grey;
-    }
-  }
-}
 
 SignalLevel signalLevelForRssi(int rssi) {
   if (rssi == null) return SignalLevel.unknown;
